% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_CI.R
\name{bin_CI}
\alias{bin_CI}
\title{Reshape and get binomial confidence intervals}
\usage{
bin_CI(data, conf)
}
\arguments{
\item{data}{The data frame generated from using data_prep()}

\item{conf}{The degree to which the Wilson confidence intervals
should be calculated, for 95\% CI 0.05 should be entered}
}
\value{
The reshaped data file with confidence intervals
}
\description{
This function take the prepared imported data (from data_prep())
then reshapes the data for use within the line graph functions,
and generates the confidence intervals.
}
\examples{
citable <- bin_CI(data=new_AMRseries, conf=0.05)
}
