% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_import}
\alias{data_import}
\title{Imports the data in the form of the AMRapp template}
\usage{
data_import(location)
}
\arguments{
\item{location}{The directory and file name of the file for import}
}
\value{
The imported data
}
\description{
This function imports the data from a pre-prepared CSV file
}
\examples{
AMR_data <- data_import("Amrapp_template_testdata.csv")
}
