% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Converting and grouping imported data}
\usage{
data_prep(data)
}
\arguments{
\item{data}{The data imported using data_import().}
}
\value{
The summarised data
}
\description{
This function takes the input file and converts dates into date format
and created a result factor. Data are then summarised in to
year and month totals at the pt_id level for genus and antimicrobial
}
\examples{
new_AMRseries <- data_prep(data=AMR_data)
}
