% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcentR_ci_plot.R
\name{pcentR_ci_plot}
\alias{pcentR_ci_plot}
\title{Plots the percentage resistance line trend, with CI}
\usage{
pcentR_ci_plot(data, filter1, filter2)
}
\arguments{
\item{data}{The data generated from the bin_CI() function}

\item{filter1}{The genus selected to plot in string format "genus"}

\item{filter2}{The antimicrobial selected to plot in string format "antimicrobial"}
}
\value{
The graph object
}
\description{
This function creates the percentage resistance line graph with
95\% confidence intervals (area)
}
\examples{
reslinegraphCI <- percentR_ci_plot(citable, "Klebsiella", "CIP")
}
