% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcentR_plot.R
\name{pcentR_plot}
\alias{pcentR_plot}
\title{Plots the percentage resistance line trend, without CI}
\usage{
pcentR_plot(data, filter1, filter2)
}
\arguments{
\item{data}{The data generated from the bin_CI() function}

\item{filter1}{The genus selected to plot in string format "genus"}

\item{filter2}{The antimicrobial selected to plot in string format "antimicrobial"}
}
\value{
The graph object
}
\description{
This function creates the percentage resistance line graph
}
\examples{
reslinegraph <- percentR_plot(citable, "Klebsiella", "CIP")
}
