% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcentRandT_plot.R
\name{pcentRandT_plot}
\alias{pcentRandT_plot}
\title{Plots the percentage resistance and tested line trend}
\usage{
pcentRandT_plot(data, filter1, filter2)
}
\arguments{
\item{data}{The data generated from the bin_CI() function}

\item{filter1}{The genus selected to plot in string format "genus"}

\item{filter2}{The antimicrobial selected to plot in string format "antimicrobial"}
}
\value{
The graph object
}
\description{
This function creates the percentage resistance and
percentage tested line graph
}
\examples{
Testlinegraph_noCI <- pcentRandT_plot(citable, "Klebsiella", "CIP")
}
