% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcentT_plot.R
\name{pcentT_plot}
\alias{pcentT_plot}
\title{Plots the percentage tested line trend, without CI}
\usage{
pcentT_plot(data, filter1, filter2)
}
\arguments{
\item{data}{The data generated from the bin_CI() function}

\item{filter1}{The genus selected to plot in string format "genus"}

\item{filter2}{The antimicrobial selected to plot in string format "antimicrobial"}
}
\value{
The graph object
}
\description{
This function creates the percentage tested line graph
}
\examples{
Testlinegraph_noCI <- pcentT_plot(citable, "Klebsiella", "CIP")
}
