% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_packages.R
\name{required_packages}
\alias{required_packages}
\title{Checks, installs and loads required packages}
\usage{
required_packages()
}
\arguments{
\item{NULL}{No parameter is required}
}
\value{
The loaded packages
}
\description{
This function checks currently installed and loaded packages
and loads those required that are not currently loaded
}
\examples{
required_packages()
}
