% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollpcentRandT_plot.R
\name{rollpcentRandT_plot}
\alias{rollpcentRandT_plot}
\title{Plots the rolling average percentage resistance and tested line trend}
\usage{
rollpcentRandT_plot(data, filter1, filter2)
}
\arguments{
\item{data}{The data generated from the roll_av() function}

\item{filter1}{The genus selected to plot in string format "genus"}

\item{filter2}{The antimicrobial selected to plot in string format "antimicrobial"}
}
\value{
The graph object
}
\description{
This function creates the rolling average percentage resistance and
percentage tested line graph
}
\examples{
Roll_av_lineGraph  <- rollpcentRandT_plot(roll_av_data, "Klebsiella", "CIP")
}
