% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_plot_function.R
\name{stack_plot_function}
\alias{stack_plot_function}
\title{Plots the incidence for each susceptibility result}
\usage{
stack_plot_function(data, filter1, filter2)
}
\arguments{
\item{data}{The data generated from the data_import() function}

\item{filter1}{The genus selected to plot in string format "genus"}

\item{filter2}{The antimicrobial selected to plot in string format "antimicrobial"}
}
\value{
The graph object
}
\description{
This function creates stacked bar plot of incidence for each susceptibility
result over time - for the selected genus and antimicrobial
}
\examples{
Stackgraph  <- stack_plot_function(AMR_data, "Klebsiella", "CIP")
}
